Locks 'thor' gem to 19.1 to workaround thor issue #538
Thor versions 19.2-19.4 generate these warning messages:

  Expected string default value for '--helper'; got true (boolean)
  Expected string default value for '--assets'; got true (boolean)
  Expected string default value for '--decorator'; got true (boolean)
  Expected string default value for '--decorator'; got true (boolean)
  Expected string default value for '--serializer'; got true (boolean)

Thor is used in rails and jquery-ui, among other gems.  None of them
locks the version.

This commit requests at least version 19.1 and refuses to load .2, .3 or
.4.  The assumption is that this will be fixed in the next release and
this fix should only prove a problem if it is not AND some other gem is
updated.